// webpack---_wix_analytics_ng_bm/constants/enums.ts
export enum CustomizationModalSaveSourceBI {
  card = 'card',
  tooltip = 'tooltip',
}
// ...
export enum DashboardWidgetType {
  horizontal = 'horizontal',
  vertical = 'vertical',
}

export enum PERMISSIONS {
  UPGRADE_SITE = 'my-account.upgrade-site',
}